requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.4
et_xmlfile==2.0.0
idna==3.11
minio==7.2.16
numpy==2.3.3
openpyxl==3.1.5
pandas==2.3.2
py4j==0.10.9.9
pycparser==2.23
pycryptodome==3.23.0
pyspark==4.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-fsutil==0.15.0
pytz==2025.2
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
